Update tests so they don't halt the JVM when running in a test.
Up until now, tests were run with a `SecurityManager` that intercepted calls to `Runtime.halt` and threw a `SecurityException` instead. But `SecurityManager` functionality has been removed from the latest JDK. So instead, the code here detects when it is running in a test and itself throws `SecurityException` instead of calling `Runtime.halt`.

PiperOrigin-RevId: 698389061
Change-Id: I74f72ab33a063f12ea843037546ef621a7afa064