BazelTestRunner: exit with 137 on OOMs
Motivation: Make it easier to detect whether a test failed due to an
OutOfMemoryException

Current behavior:
The test runner exited with a code 1 on any test failure.
This can make it hard to programmatically differentiate whether a test
failed due to an actual test failure, or because the test was aborted
due to an OutOfMemoryException.

Proposed change:
If a run of a test suite fails, the failures are checked for any OOMs.
If at least one failure was caused by an OOM, the program exits with a
code 137, rather than 1. Additionally, this is also the exit code if
the test runner itself OOMs.

This distinction can help with automatically detecting OOMs and then
retrying the execution with more memory available, and is especially
helpful in the context of remote execution, where the action may be
retried on a larger executor.