[8.0.0] Re-enable symbolic macro attribute inheritance and remove --experimental_enable_macro_inherit_attrs flag (#24445)
After
https://github.com/bazelbuild/bazel/commit/a2f1f58cdff07d240c9f715d5091672bd9abd03f,
macro inheritance API is in a good shape - and without macro
inheritance, we wouldn't have any good way for symbolic macros in Bazel
8 to use `testonly`.

Considering for how little time the
--experimental_enable_macro_inherit_attrs existed, and that no
references to it was marked experimental, it seems safe to remove it
directly without graveyarding.

Cherry-pick of

https://github.com/bazelbuild/bazel/commit/a39d18122ae824697691813a103ccb65871ff033

RELNOTES: Re-enable symbolic macro attribute inheritance.
PiperOrigin-RevId: 698884636
Change-Id: I7f682949e96b6739e12d8f569a87ef2daf0ea41c