Reduce verbosity of BulkTransferException
BulkTransferException was sometimes overly verbose in the console, particularly when many transfers failed for the same reason.

This commit shortens and normalizes BulkTransferException messages to improve the efficiency of existing event deduplication, such as that performed by _RemoteExecutionService.report()_.

Normalization involves not mentioning individual digests in exception messages from ByteStreamUploader. Instead, we now wrap StatusRuntimeException directly in an IOException, which is a common pattern in the GrpcCacheClient.

Closes #23490.

PiperOrigin-RevId: 699081591
Change-Id: I1d02e19200e4a84b7c36813a1705fb039a1f60a9