Take a shared lock on the install base from the client side.
These are the client-side changes required to implement garbage collection of stale install bases: since one Bazel server might attempt to collect the install base of another, we must ensure that a running Bazel server *or* client prevents collection of its own install base, which is achieved by acquiring an exclusive lock prior to collection (to be implemented in a followup).

Note that we keep the existing mechanism for handling concurrent attempts to create the same install base don't clash (atomic filesystem move). We can't use an exclusive lock to synchronize creation because, when the install base already exists, it would prevent concurrent invocations.

PiperOrigin-RevId: 699165035
Change-Id: Iac3273d6548f59c31bb8fa96288d34e69f222102