Automated rollback of commit aa7a577ac8dde4b692d76cbd994516cc91987c08.
*** Reason for rollback ***

Breaks blaze postsubmits.

*** Original change description ***

Take a shared lock on the install base from the client side.

These are the client-side changes required to implement garbage collection of stale install bases: since one Bazel server might attempt to collect the install base of another, we must ensure that a running Bazel server *or* client prevents collection of its own install base, which is achieved by acquiring an exclusive lock prior to collection (to be implemented in a followup).

Note that we keep the existing mechanism for handling con...

PiperOrigin-RevId: 699218665
Change-Id: Id252767635132ca7e95a00523adc5786ae08726f