Refactor the unit tests to allow us to retrieve the ExecutionPolicy of the post-processing spawn even if the spawn is not executable in local.
In an upcoming refactor, we'd like to provide more informative log/exception messages to let users know what's wrong behind a failed execution.

In order to create such messages, it will be better if we can retrieve the execution policy before we check if either spawn is executable. That is because we want to embed the information from the execution policy to the log/exception messages. To enable that, we need this refactor in the unit tests.

RELNOTE: None
PiperOrigin-RevId: 700293992
Change-Id: Ic5ca6b82049d5001d652d551dab388f14fcf4577