Call native collect_libraries_to_link
The Starlark implementation causes a regression, because it's running a long for loop.

Pass supportsDynamicLinker instead featureConfiguration into convertLibraryToLinkListToLinkerInputList. This way the native and Starlark interfaces are the same.

Expose the interface from cc_internal and call it from Starlark.

There's one more list copying because calling native interface, but it's not measurable.

PiperOrigin-RevId: 700328659
Change-Id: I37a6b73d174889119168d1acfb8f0cb612b6542d