Add query option proto:include_starlark_rule_env
If set to false, this will omit starlark rule environment in a rule target's proto representation. This may be useful to isolate the representation for only properties which are visible in a query environment. For example, if set to false, the starlark rule implementation function of a target does not contribute to its proto representation in any way, though its attribute values and rule name do.

PiperOrigin-RevId: 700732792
Change-Id: Ic77b017838dfe26a2f454634156be8e554818e44