Update visibility section on macro page
This is a rewrite to present the most relevant information, while eliding some technical detail. In particular, it reduces mention of visibility delegation to an afterthought, and focuses instead on the key insight and a single demonstrative example.

A subsequent CL will update visibility.md itself with appropriate details.

Work toward #23857.

PiperOrigin-RevId: 700736327
Change-Id: I543b57746e00bd9e62adb2b6d50ea25abc6e756f