Re-enable testMultiplexWorkRequest_stopsWorkerOnException which depended on GoogleTestSecurityManager but no longer does.
We let the second work request initiate an interrupt to the reader thread in the MessageProcessor. This will then cause the PipedInputStream under the hood to throw an InterruptedIOException. This process helps us simulate the situation when the infinite loop in the WorkRequestHandler catches an IOException while calling `messageProcess.readWorkRequest()`, then it'll trigger the following path of stopping the worker on an exception which we're testing.

RELNOTE: None
PiperOrigin-RevId: 701200352
Change-Id: I77a8d06a74f519ea347a99e5fecb81c79aaf9713
(cherry picked from commit e53ed1df40b0233daf5be5b4f2f4a54391bbfb7c)