Use fixed filenames for the persistent action cache.
The file contents are already versioned, with a version mismatch on load causing them to be saved under a different name, logged and ignored. There's no correctness need to also bump the filename on format changes.

This simplifies the task of garbage collecting the on-disk representation, because we don't have to look for files left over by format changes.

One adverse consequence is that switching back and forth between Bazel binaries built before and after a format change will now invalidate the cache, but format changes are infrequent enough (less than 1 per year) for this to be a concern.

PiperOrigin-RevId: 702316556
Change-Id: Ic8d78b2bcf6ef3e0b40777d65ca8f96b27583db1