Remove kythe task from presubmit.
Sometime it can be the bottleneck when other tasks are mostly cached.

I don't think kythe task in presubmit provides much value: they are built with `--keep_going` so the task is green even there are errors. People will probably never check the logs.

In case of breakage, we will get notified by the internal pipeline.

PiperOrigin-RevId: 702355117
Change-Id: I534603993949824e97045a8a7a84ad0b30066492