Handle changed files outside of active directories for remote analysis caching.
This is a _coarse grained and conservative_ approach to disallow clients from using remote analysis caching when there are changed files outside of the active directories. This is because such files __may__ invalidate a deserialized frontier, but is unable to due to the pruned Skyframe graph, and ultimately cause [stable inconsistent builds](https://bazel.build/run/build#build-consistency)

This is implemented with the `--experimental_frontier_violation_check` flag and only triggers with `--experimental_remote_analysis_cache_mode=download`.

The flag has 3 strictness levels:

1) `strict`: **preferable if reverting the local changes is manageable**. Stop the build immediately and report which filesystem changes are affecting the build. Progress can only be made when the violating change is reverted, or with the `warn` option.
2) `warn`: **preferable if reverting the local changes is undesirable**. Gracefully fallback to fully local evaluation by disabling analysis cache downloads, while warning the user which files are affecting the build. If there exists at least one remotely fetched analysis value in the Skyframe graph already, first invalidate all existing analysis and execution phase values, then continue the build without remote analysis caching.
3) `disabled_for_testing`: do nothing. This is a test specific escape hatch.

PiperOrigin-RevId: 702986396
Change-Id: Ia031e57810bac8fae817d88640c44a4d1c47f307