Add a shutdown hook for tests that detects `System.exit`.
We can no longer prevent code being tested from exiting when `System.exit` is called, but we can at least show where it happened. We do this by the examining the stack traces of all threads in a shutdown hook, looking for `Runtime.exit`. (`System.exit` calls `Runtime.exit`, and users can also call it directly.)

We cannot detect `Runtime.halt` this way, unfortunately.

Recent JDK versions can be configured to log where `System.exit` is called, but that logging is fiddly and its configuration may interfere with that of the code under test. Also, we may later add to the shutdown hook so that it also detects which test method is currently executing, in case that isn't in the thread that called `System.exit`.

PiperOrigin-RevId: 703190140
Change-Id: Ie4ba4728c41adb868f7fc75b630851b900e7e27d