xml test output: support scientific notation without decimal
Boost test when writing XML files may sometimes omit the decimal point,
but still write in scientific notation as floating point seconds. To
support reading these XML files, if the string contains an `e`, then
try to parse as a float in hopes it's a floating point number. Since
parseLong will not support a `e` character this seems like a safe change
that is unlikely to break anything.

Note that the comment here is fixed as it was phrased opposite of the
actual implementation.