Fix infinite recursion when querying a target whose visibility is defined by a circular package_group
Such a target can't be built, but we still shouldn't crash when we attempt to run `bazel query` on it (using the `visible()` operator). This CL detects the infinite loop and prunes the traversal, yielding a valid query result on a target that still can't be built.

Theoretically this may also reduce traversals when the `package_group` `includes` graph is a DAG, but that effect won't be noticeable given how rare deep `includes` are.

PiperOrigin-RevId: 704305833
Change-Id: I4277077f85f8fd5c57d78218db982e40fa8a1299