Make the llvm coverage tests a little more robust.
Check that the actual coverage data simply contains the expected
coverage data, rather than checking for an exact match. This matters for
when we start to process baseline coverage, which may result in header
files getting included in the report. This is in-keeping with the other
coverage tests.

Also adds the branch data into the reports - this is never disabled when
using LLVM so we should check it.