[7.6.0] Make the llvm coverage tests a little more robust.
Check that the actual coverage data simply contains the expected
coverage data, rather than checking for an exact match. This matters
for when we start to process baseline coverage, which may result in
header files getting included in the report. This is in-keeping with
the other coverage tests.

Also adds the branch data into the reports, which we will check when
we're using an LLVM version that supports branch coverage, and a test
that checks header files that include coverage data are actually
included in the final report.

Required for https://github.com/bazelbuild/bazel/pull/24593

Closes #24624.

PiperOrigin-RevId: 705177531
Change-Id: I2247434d0b005f2218761249238e52597a452e12

Commit: 164c4821451c8b03a2e4560ac85c04ae2fe8462e