PROJECT.scl files can set an enforcement_policy indicating how/whether they allow output-affecting flags in the command line or user RC.
enforcement_policy=warn
Warn if the user sets an output-affecting flag in the command line or user RC.

enforcement_policy=compatible
Fail if the user sets an output-affecting flag to a value that conflicts with the selected config.

enforcement_policy=strict
Fail if the user sets any output-affecting flags not already set by the applied config.
PiperOrigin-RevId: 705246363
Change-Id: I453c527bb62849b22642849d82c66bec097b9675