Clear the credentials cache whenever `--credential_helper_cache_duration` changes.
This provides a convenient way to clear the cache without running a clean command. While we could update the expiration on existing entries, Caffeine doesn't make it easy and it's unclear that the added complexity is worth the benefit (most users will pick an appropriate value once and stick with it).

Also upgrade Caffeine so that tests can use a recently added API.

Fixes #24437.

PiperOrigin-RevId: 705426812
Change-Id: I4b6d055bd0cda847a5bfd678f38aac705e2692b1