Fix symlink planting for multiple aspects
When Skymeld is enabled, symlinks for external repositories are planted lazily in response to an event sent in `BuildDriverFunction`. For top-level aspects, this event needs to be sent individually for each aspect as the transitive packages required can differ per aspect. `BuildDriverFunction` deduplicated all events, including the one for symlink planting, by type, which resulted in missing symlinks for actions registered by the second (and later) aspects.

Fixes #24619

Closes #24676.

PiperOrigin-RevId: 707458188
Change-Id: I9fe523bc746d29e6c6ddb6367e08ce2c6c48c69b