Update Bazel's embedded JDK to 23.0.1
In addition to performance improvements, this JDK release contains a fix for https://bugs.openjdk.org/browse/JDK-8330077, which allows `--watchfs` to handle a configurable number of changes to a watched path rather than maxing out at 500.

Windows arm64 can switch to Zulu JDK 21.

Since async-profiler is released very rarely (last release in January) and the latest release is not compatible with JDK 23, this PR consumes a recent nightly release instead. This has the side effect of removing a bit of bloat from the server jar as it no longer contains async profiler's native library for all OSes/arches.

Closes #24609.

PiperOrigin-RevId: 707717223
Change-Id: I29fe86eca1e0deeae813e7dbbf7d60859e3cba5e