Check for specifically test trimmed options in `config_setting`.
If a `config_setting` refers to a test option when the test configuration has been trimmed, consider that option to not match. Note that this specifically does not check whether the expected value is the default: it will still fail to match because the option is not present.

Work towards merging #24398.

PiperOrigin-RevId: 712604907
Change-Id: I2afbbe82de091770789988316a6c450c1fdc8f39