[7.5.0] Include apparent names of deps in `bazel mod` JSON output (#24847)
This allows IDEs to query for the direct dependencies of the root module
as well as how they can refer to them from the point of view of the root
module.

Also always emit `name` and `version` so that consumers don't have to
know how to parse module keys.

Work towards #22691

Closes #23787.

PiperOrigin-RevId: 693453084
Change-Id: Ie3fd5e89301d8e83d0eaa686188634923853f01a (cherry picked from
commit 1003d2c31bf4f90ded7015b1b7ed570d50ed6a63)

Fixes #24678