[7.5.0] Use long instead of int for thread indices (#24866)
This protects against an integer overflow which could occur for large
key list size and large thread counts.

Regrettably, it's difficult to write a regression test for this
scenario, as exercising this overflow requires lots of time and heap, so
it would be a performance regression to our test suites.

Fixes https://github.com/bazelbuild/bazel/issues/19445

PiperOrigin-RevId: 595420516
Change-Id: Ic0a475a6a273c50fe9895dd0852fa5b062859cb2

Commit
https://github.com/bazelbuild/bazel/commit/3e373d0abc79e40b3fa46aca4aa268c5ae9b7dc2

Co-authored-by: Googler <cparsons@google.com>