Respect comprehension variable shadowing in Starlark debugger output
The Starlark interpreter, for optimization reasons, stores the bindings of
comprehension variables in their enclosing function's locals array. This
means we cannot naively transform the locals array into a name -> value map
for debugger output: there may be multiple initialized variables having the
same name.

And since we have been building this map using buildOrThrow(), attempting
to debug a function with a comprehension variable having the same name as
a parameter or other local variable would crash Bazel.

The immediate fix would be to track comprehension bindings' lexical scope
(by saving a pointer to the comprehension AST node; note that a
comprehension's scope has a non-trivial shape - it has a "hole" punched out
for the first `for` clause), so we can ignore comprehension variables outside
their comprehension, and emit only the innermost one when inside a
comprehension (relying on the fact that the resolver places inner
comprehensions' variable bindings after the outer ones).

However, while this fixes the crash and the most obvious correctness problem,
it does not allow the user of the debugger to examine the value of the
shadowed variables.

The proper fix - to be implemented later - would be to push a new debugger
frame when debugging inside a comprehension.

See https://github.com/bazelbuild/bazel/issues/24931 and discussion with @fmeum in https://github.com/bazelbuild/bazel/pull/24919

PiperOrigin-RevId: 715908968
Change-Id: I3066e89f2e92d0fd0e483851a767b7c7d70ab555