Fix performance regression in CriticalPathComputer.
Previously, when actions are change pruned, we add their entire transitive closure to critical path graph. This could slow down incremental builds if the action graph is large and only a small portion of it is invalidated.

This CL improves that by only adding actions that are invalidated (including change pruned).

We achieve that by updating Skyframe to emit change pruning event. These event are only emitted for nodes that were invalidated by dirtiness check during incremental builds. So that the CriticalPathComputer now operates on a subset of action graph that only contains dirty actions. This is more correct and performant for incremental builds.

PiperOrigin-RevId: 716146196
Change-Id: I71add4240458b5d4087b1af81a41202f98fee947