Add `--incompatible_enforce_utf8`
If enabled (or set to 'error'), fail if Starlark files are not UTF-8 encoded. If set to 'warning', emits a warning instead.

Bazel already assumes that Starlark files are UTF-8 encoded for e.g. filenames in actions executed remotely. This flag doesn't affect this, it only makes encoding failures more visible.