[7.5.0] Do not invalidate remote metadata during action dirtiness check (#24941)
... if the corresponding output is materialized afterwards and hasn't
been changed.

Failing to do so will cause an incremental build to re-check the action
cache for actions whose outputs were materialized during last build.
This can be very slow if the size of the outputs are large.

We achieved this by storing `FileContentsProxy` in the remote metadata
after materialization. During dirtiness check, we compare remote
metadata and the corresponding local metadata with their `digest`, or
`proxy` if `digest` is not available for local metadata,

A user reported back the wall time of their very first increment build
is improved by this change from `14s` to less than `1s` in common case,
and from `115s` to less than `1s` in worst case.
https://github.com/bazelbuild/bazel/issues/24763

PiperOrigin-RevId: 715734718
Change-Id: Id1a1a59d8b5f3e91a7ae05a73663ff37eee6d163

Fixes #24920.