Fix BuildEventStreamer to atomically post `Configuration` events while recording the configuration posted.
In the Build Event Protocol, any number of events may reference a single
configuration. For example, every `TargetComplete` event will reference its
configuration and most builds will have many targets and only a small handful of
configurations.

To simplify BEP consumption we aim to report an event's Configuration before we
report the event itself. We also do not want to report a copy of the same
Configuration for every single event that references it. So we only post the
configuration the first time we see it referenced.

We need to combine "check if this is the first time we saw the configuration,
recording if it is" with "actually post this configuration" in a single atomic
unit. Otherwise, when two targets finish in separate threads and reference the
same configuration, a race occurs.

RELNOTES: n/a
PiperOrigin-RevId: 716809640
Change-Id: I559a1d85e06d82144f98f24681ee357704ef5f5c