Unify implementations of `Spawn#getCombinedExecProperties`
`SimpleSpawn` returned the exec properties of the action owning the spawn, while `BaseSpawn` returned the exec properties of the exec group owning the action that owns the spawn. The latter contains the `exec_properties` set on the target or exec platform, whereas the former is entirely specified by the action.