Fold `processResult` into `processErrors` in `SkyframeErrorProcessor`.
These are 2 different ways that can handle execution phase errors.
This CL would simplify the code and remove code divergence.

There should be feature parity between the 2 paths as in fact the `processErrors` path has been used for Skymeld for a while. However, as this touches the inner workings of Skyframe, it will be rolled out with a flag to be on the safe side.

PiperOrigin-RevId: 717476404
Change-Id: Iba0aa358bae3f8ecb75702e78228cb74a3edd663