Reduce garbage created by `LtoBackendAction#computeBitcodeInputs`.
In a build with many `LtoBackendAction` this CL reduces:
* Garbage allocated by that method by 99.8%
* Total time spent doing full GC pauses by 97.3%
* Wall time by 74.3%

Don't pay too much attention to the exact numbers. I did only a single trial since each trial takes ~40m. Instead, just pay attention to the direction of the change :)

Shout-out to jhorvitz@ for pointing out that `BitcodeFiles` is shared among `LtoBackendAction` instances and so weakly caching the result is the biggest macro optimization we can do.

PiperOrigin-RevId: 718087528
Change-Id: I017528e60b1d23e430f1a5d06bcfa8061092057c