Avoid explicit references to RemoteFileArtifactValue.
Instead, pass a FileArtifactValue around and use isRemote() and its related accessors when required. In particular, OutputChecker (previously RemoteArtifactChecker) now receives a FileArtifactValue and must handle non-remote metadata appropriately.

This makes the FileArtifactValue class hierarchy easier to change. Specifically, I'm planning to extend the concept of a materialization path to non-remote artifacts, and the most convenient way to do that is to introduce a subclass that wraps around any of the other subclasses, which will cause instanceof checks to stop working.

PiperOrigin-RevId: 718851433
Change-Id: I6369fc3dbdc3a93380a0a1875b631216c0eed3dd