Avoid explicit references to InlineFileArtifactValue.
Instead, introduce an isInline() method.

This makes the FileArtifactValue class hierarchy easier to change. Specifically, I'm planning to extend the concept of a materialization path to non-remote artifacts, and the most convenient way to do that is to introduce a subclass that wraps around any of the other subclasses, which will cause instanceof checks to stop working.

PiperOrigin-RevId: 718897103
Change-Id: I2fdfaa55d49cbd94a9b53c36450fc660bf37b241