Update FileSystemDependencies match methods to use VersionedChanges.
Renames containsMatch to findEarliestMatch, which is more consistent
with it returning the matching version instead of a boolean. Also passes
a validityHorizon parameter.

Also splits out FileOpDependency from FileSystemDependencies
to allow specifying dependency sets excluding
NestedDependencies. NestedDependencies are treated differently downstream
due to the separation of analysis and execution dependencies.

PiperOrigin-RevId: 718931169
Change-Id: I4501b7dc9d77c9dd898a1bd06e87c291bb0cf23c