[8.1.0] Reduce retention and early materialization of `VirtualActionInput` (#25008)
https://github.com/bazelbuild/bazel/commit/b82dcdcfe71e8160c37a0963a5ed1b586e819266
introduced a way for parameter files to be materialized lazily and in a
streaming fashion. For this to be effective, the contents of
`VirtualActionInput`s should not be retained and, ideally, never or only
briefly materialized in full.

This is achieved by replacing the digest computations with a streaming
approach and avoiding retention of `VirtualActionInput` in
`MerkleTree`s, which may be cached. All cache protocols are streamlined
to accept a `Blob`, which is a closeable supplier of `InputStream`s that
is automatically closed after the upload has completed (which may
require retries).

Along the way, file nodes in `MerkleTree`s are stored unwrapped to
reduce memory usage.

Closes #24891.

PiperOrigin-RevId: 718042524
Change-Id: I1567cf6ca36cbf5e1146c91df34157c9192bb822

Co-authored-by: Fabian Meumertzheim <fabian@meumertzhe.im>