PROJECT.scl + expanded flags warns with original flag (eg --config), once, instead of for each output-affecting flag in the expansion.
I'm on the fence about this behavior because for --configs that contain just a single output-affecting flag amongst a bunch of flags that are allowed to be set alongside the --scl_config, this change makes it more difficult to see why the --config isn't allowed. The pro is that for configs that set a lot of output-affecting flags, there's a lot of log spam before this change.

PiperOrigin-RevId: 719407331
Change-Id: Ic51c54c46f58fbca206b04e57047f92c3d6db5f7