Fix Aspect on Alias Configuration Key
For alias chain, i.e if one alias target has another alias as its `actual`, the configuration of the `actual` target can be different from the parent alias in case of flags scoping. This change uses the configuration of the `actual` target in the `AspectKey` created for it to avoid
`Aspect not in same configuration as base configured target` failures.

PiperOrigin-RevId: 719418094
Change-Id: I467101f815e1b230627886100fa3a6c923f4fb44