[8.1.0] Distinguish between input and output prefetching in request m… (#25058)
…etadata

ea4ad30d4e49b83b62df0e10b5abe2faadea7582 added an action as context for
prefetches with BwoB, but assumed that all callers would use the
prefetcher for inputs to locally executed actions. However, since then,
it has also been used to download outputs of completed actions that have
been explicitly requested (e.g. as outputs of top-level targets or those
matching the regex path patterns). This resulted in downloads with a
`prefetcher` action ID and action details for the action that produced
the file, rather than consumed it, resulting in confusing situations for
observability tools.

This is fixed by separately tracking the reason for the fetch. Using an
action ID of `output` when the action has the requested file as an
output, and "input" when the action has the requested file as an input.

Closes #25040.

PiperOrigin-RevId: 719246746
Change-Id: Ib95ff65ba68112b1a38ab3022e5b1a19ef74cc9f (cherry picked from
commit 998e7624093422bee06e65965f8a575d05d57c27)

RELNOTES: CAS requests made when Bazel downloads a blob with Build
without the Bytes enabled now provide metadata with an action ID of
input if the blob is downloaded as the input to a local action and
output if it is a requested action output.