Check `future.isDone()` before processing it.
In unknown commit, the value type of the map was changed from `Collection<Inclusion>` to `ListenableFuture<Collection<Inclusion>>` but the check didn't change. It means we could await a pending future on current thread, instead of dispatch it to the thread pool, which is the code did before unknown commit. This CL makes the cache work as intended.

PiperOrigin-RevId: 720124092
Change-Id: I9e05cfe2840237aad8c5872e7c0693e5db2ed124