[8.1.0] Roll forward of https://github.com/bazelbuild/bazel/commit/aa7a577ac8dde4b692d76cbd994516cc91987c08: Take a shared lock on the install base from the client side. (#25085)
These are the client-side changes required to implement garbage
collection of stale install bases: since one Bazel server might attempt
to collect the install base of another, we must ensure that a running
Bazel server *or* client prevents collection of its own install base,
which is achieved by acquiring an exclusive lock prior to collection (to
be implemented in a followup).

Note that we keep the existing mechanism for handling concurrent
attempts to create the same install base don't clash (atomic rename)
because it's simpler than using the lock (which would require upgrading
it from shared to exclusive and back).

Progress on #2109.

PiperOrigin-RevId: 702433270
Change-Id: I474f3d56ec126de5f975c543f7bf9b64f4f08124