cc_binary: implement `additional_link_outputs` feature.
The feature allows for retrieving additional files created by the link
command.

The additional filenames are determined by the `cc_toolchain_config`
attribute `additional_link_outputs` which lists the suffixes that will
be added to the binary name.

For binary `out` where `additional_link_outputs = [".a", ".b"]`, we'll
expect the following files to be created by the link action:
- `out`
- `out.a`
- `out.b`

Solves #24990.