Avoid explicit references to UnresolvedSymlinkArtifactValue.
Instead, use getType().equals(SYMLINK) where needed, and move the getUnresolvedSymlinkTarget() method (previously getSymlinkTarget()) to the base class.

This makes the FileArtifactValue class hierarchy easier to change. Specifically, I'm planning to extend the concept of a materialization path to non-remote artifacts, and the most convenient way to do that is to introduce a subclass that wraps around any of the other subclasses, which will cause instanceof checks to stop working.

PiperOrigin-RevId: 720570652
Change-Id: I614b666e8ab18b5e9b830f006f5c1a2c1ea4e66e