Remove the "overwrote runfile" warning.
At HEAD, this is reported on the console as an error but doesn't actually cause a build error and this has been the case for a long time. This means that this particular cat is out of the bag and we can't prohibit this behavior anymore and therefore it's better to not emit spam.

RELNOTES: None.
PiperOrigin-RevId: 720983277
Change-Id: I2f62247b7bb7a81e30281ab578a8b93f17cef476