Removes the `--incompatible_enable_cc_test_feature` flag.
This flag switches from the use of build variables to the feature of the same name.  The functionality this was intended for was never completed, and is no longer needed, so we are rolling things back.

RELNOTES:
Removes the `--incompatible_enable_cc_test_feature` flag.  The functionality this was intended for was never completed, and is no longer needed.
PiperOrigin-RevId: 721682446
Change-Id: I3e3c7b1042542e2ce8042b981e1302c5bcf12340