Don't let the `test` exec group inherit from the default exec group
Part of #24964.

RELNOTES[INC]: Constraints and toolchain requirements added to the default exec group, for example via the toolchains parameter of the rule function or the exec_compatible_with attribute on all rules, no longer apply to the test exec group, which contains the test action on test rules. Instead, use exec_group_compatible_with to apply constraints and/or define an explicit test exec group with toolchain requirements on test rules that require it.

PiperOrigin-RevId: 721834175
Change-Id: I8504ff1a08a643ed35a7282f09f6d6faf75d574c