Generalize the concept of a FileArtifactValue with a resolved path.
The presence of a resolved path is now orthogonal to the remainder of the metadata, i.e., any existing metadata may be wrapped with a resolved path. The added flexibility will be useful in upcoming changes extending the concept of lazy materialization to non-remote artifacts.

This CL doesn't change preexisting behavior; it's just an API refactoring.

PiperOrigin-RevId: 722116113
Change-Id: I85234bf7363ea3361ff756bd65015010fff618a9