Extend `BazelFlagInfo` with additional metadata
This commit adds additional metadata to the protobuf dump produced by
`bazel help flag-as-proto`.

The old name and the deprecation message are useful to provide
diagnostics within bazelrc files using a language server.

The `type_converter` and `enum_values` are usful to validate argument
values and to provide auto-completion.

The option expansions can be used, e.g., to warn on contradictions
between explicitly specificed flags and expansion flags.