Fix BEP inlined outputs for aspect-complete events when outputs must be uploaded.
The recently introduced option `--experimental_build_event_output_group_mode` allows users to request certain output groups to have their `File`s reported inline in the `TargetComplete` event. This allows users to more easily obtain URIs for files without walking the `NamedSetOfFiles` events, while accepting the responsibility to ensure that the requested output groups are small enough that the `TargetComplete` event does not grow enormous. (Unlike `NamedSetOfFiles`, a `TargetComplete` event cannot be split up recursively.)

The default BEP behavior is to only report a `File` with a remote URI. In some build configurations, outputs might already have a remote URI determined during execution, but this is not true in general. When an output file that appears in BEP does not have a known remote URI, the BEP subsystem is responsible for uploading the file to obtain a remote URI.

The `NamedSetOfFiles` event has working uploading logic, so files were correctly reported by aspects using the default value of `--experimental_build_event_output_group_mode`. Previously, the uploading logic was missing for `AspectCompleteEvent`, so we did not have a remote URI when inlining was requested by `--experimental_build_event_output_group_mode`. This change rectifies this oversight, ensuring a remote URI is available no matter how an aspect-produced file is reported.

PiperOrigin-RevId: 722658204
Change-Id: Ie6fc70192f687b8e5582d76ab048dfa6bd25186e